#5: Let dataController show real values

+++ src/main/java/de/aikiit/bilanzanalyser/api/DataItem.java
@@ -1,4 +1,4 @@
 package de.aikiit.bilanzanalyser.api;
 
-public record DataItem(String name, int value) {
+public record DataItem(String name, long value) {
 }
